[hydrostatic] Fixed bug discovered implementing 6DOF
Shouldn't use idxA & idxB to generate the output polygon:
those are transformed indexes in that they refer to the
indices of the points in the original mesh, whereas first_and_last.first
and first_and_last.second refer to indices in the facet
under consideration. Use idxA & idxB to compute idxA1, idxB1, P
and Q but not to loop on the nodes of a facet.