[hydrostatic] Added -Wconversion flag
Removed all implicit casts. This was done because a bug was
discovered in the hydrostatic module, due to implicit casting:
a string was provided in lieu of a Point object but since the
Point class has a constructor taking a string as parameter, the
string was silently converted to a Point, its coordinates being
initialized at zero.
The -Wconversion flag is supposed to stop these kinds of problems
from occruring in the future.