[oscillating_cube] Fixed bug in hydrostatic
When some nodes are exactly on the surface, the algorithm
doesn't choose the right branch: nodes on the surface
should neither contribute to the number of emerged nodes
nor to the number of immerged nodes. We needed to
distinguish between: (1) fully emerged, (2) partially
emerged & (3) fully immerged, which cannot be done
simply by counting the emerged nodes.