[improve_computation_time] Modifications to compile with CLang
Clang does not recognize tr1: either it compiles in C++11 mode
or it doesn't. Therefore, to use std::tr1::shared_ptr and
std::tr1::unordered_map, we have to go into C++11 mode and include
<memory> and <unordered_map> instead of <tr1/memory> and
<tr1/unordered_map>.
The other consequence is that we must compile everything using
the -std=gnu++0x flag.