[refactoring] Modified YAML for g & rho
Environmental constants are now in the 'environmental constants'
section of the YAML file. The previous implementation (i.e.
define these constants as part of the force model parameters)
had the advantage of handling those parameters like any other
force model parameters, but in fact those constants are *not*
like any other force model parameter because:
(1) they can be used for several force models (which means they
    would have to be re-declared for each one, possibly
    inconsistently)
(2) it's hard to imagine a ship simulation that won't be using
    them, except the falling ball example which doesn't use rho

The previous implementation also had the following disadvantages:
- Defining the constants separately for each body could lead to
  problems (eg. different g & rho for different bodies, which
  almost certainly is not what we want)
- When implementing SimulatorBuilder::get_environment_and_frames,
  we cannot fill the structure because the Body structure purposely
  does not contain the force model parameters. This stems from
  point (1) above.