[bug_immerged_polygon] Normalized quaternions
Doing it in a fairly ugly way:
- Have to copy the states because the solver expects the states
  won't be modified
- Doing it at each time step which is potentially dangerous as
  the solver is not seeing the same states as the models

A clean way to do it would be to divide the quaternions into
a continuous & a discrete part, the discrete part being periodically
set to the difference between the normalized quaternion & the
"solver" value. This way, the states seen by the solver are the
same as the ones seen by the model.