[test_cubes] Added (disabled) test for fully immerged rectangle
Shows the effects of the point of application of the hydrostatic force:
The point of application is *not* the centroid of the triangles (because
of the pressure gradient). Code should be updated so that this test
passes.
We Added this test because we realized that the "oscillating cube" test
will not pass unless the point of application of the hydrostatic forces
is corrected. This test is merely a planar simplification of the
"oscillating cube" test.