[dev] Normalized quaternions at each time step
This is usually a very bad idea because the solver calls
dx_dt = f(x) but internally x is replaced with normalized(x)
so the model doesn't see the same states as the solver which
can lead to numerical problems.
One way to do it is to write
x = xcontinuous + xdiscrete
where xcontinuous is not normalized & xdiscrete = xcontinuous0/norm(xcontinuous0) - xcontinuous0
and xdiscrete is periodically recalculated.
This implies simulating discrete models & would take us too far.