[point_application_efforts_hydrostatiques] Implement algorithm for intersecting a Mesh with free surface
The mesh is now embedding the relative immersions of each of its vertices.
It is thus able to compute the intersection with free surface by himself.
Of course, the list of relative immersions must be updated at each step (see Mesh::update_intersection_with_fee_surface())
New vertices edges and facets are created dynamically when it is necessary to split a facet into an emerged and immerged parts.
After the update, the set of emerged (immerged) facet index are available for looping only on those facets.
Code compiles but algorithm is not yet tested.