[point_application_efforts_hydrostatiques] Correct bug in Mesh when free surface crosses exactly two vertices of a Facet
In this case, no Edge intersection with free surface is detected, and the facet is not split.
That does work for triangle, but not for n-gons n>=4.
It is necessary to add a status bit indicating whether the edge potentially crosses free surface
(to be understood as: the facet containing this edge potentially crosses the free surface)