[froude_krylov] Can parse Froude-Krylov in YAML
Right now it's a quick & dirty hack: Froude-Krylov is computed as a
side-effect of the hydrostatic module, which is bad (because they
are unrelated & because it stops us from separating the two in the
output).
The Right Way (c) to do it is to use the refactoring which
is currently under way & will allow us to iterate on the mesh.