[bug_2715] Fixed bug in MeshIntersector::has
It's quite possible that the closing facet (automatically computed) doesn't
exist as a single facet, but as the union of several facets.
For instance, when a cube is meshed triangularly, the closing facet corresponds
to the union of the two (triangular) top facets, therefore we should not add it
a second time.