[html] Added default constructor to WebSocketClient
Because we need it for WebSocketServer:
The lambda must give access to the websocket (so the server can eg. send back a
reply, as eg. an echo server)
That websocket isn't buildable until the server is started but the lambda is
needed to initialize the server
Therefore, the lambda stores a pointer to an empty websocket which is
initialized once the server is up.