[dev] Fixed a bug in SurfaceElevationInterface::get_output_mesh_in_NED_frame by refactoring SurfaceElevationInterface.cpp
In get_output_mesh_in_NED_frame, a swap instruction was missing for the computed transformation.
The instructions in this function were already present in other functions.
The idea was to use a single function to perform all these tricky operations.

This template function was renamed to be more explicit
compute_relative_position -> compute_position_in_NED_frame