Set HOS parameters outside the constructors
As we got ourselves a nice singleton, we only use the constructor once, which
means the parameters cannot be changed if they are set in the constructor.
Check the set_parameters method has been called before requesting anything.
That's the downside of course: the object can now be in an invalid state...