Fixed bug 3241
The problem was that the states are forced when calculating the
derivative, but not when observing the system.
The observer uses the integrated states (which were not forced)
supplied by the solver. The solver never sees the forced states
as they are computed by Sim::operator(). The force models use
the forced states so the forces & torques are correct, but the
state values are not.
The simplest way to fix it is to force the states in Sim::output:
it's a little redundant, but computing the forced states is not
particularly resource-intensive...