:sparkles: Upgraded solver for forced states
The problem was that unforced states were being integrated, which meant
that as soon as state were "unforced" again, the simulation proceeded as
if no state forcing had ever occured. The only way to fix that was to
fix the ssc's solver, which was a bit too involved due to the current
build & deployment process. We will need to create an issue to simplify
the deployment & another to implement the fixed solver in the ssc.