:white_check_mark: Adjusted value in unit test 'LONG_can_simulate_radiation_damping'
As far as I can tell the value was arbitrary anyway since it did not
detect the wrong operation in the force model. A unit test that
highlights the issue should probably be implemented in the unit tests
directly related to the force models, but I don't know where exactly
because the force models violate the Single-Responsibility principle
which make such operations difficult to test individually.

Signed-off-by: Moran Charlou <moran.charlou@ec-nantes.fr>